update type for occurrenceids--needs big int

+++ src/graphql/schemas/plantData.schema.ts
@@ -21,7 +21,7 @@ const PlantDataCommonFields = `
   hardiness: [Int!]
   uses: [String!]
 
-  occurrenceIds: [Int!]!
+  occurrenceIds: [Float!]!
   mediaUrls: [String!]!
   scrapeSources: [String!]!
 `;
